+++ src/main/java/com/schwab/urlshortener/exception/UrlMappingNotFoundException.java
@@ -1,0 +1,8 @@
+package com.schwab.urlshortener.exception;
+
+public class UrlMappingNotFoundException extends RuntimeException {
+
+    public UrlMappingNotFoundException(String shortCode) {
+        super("URL mapping not found for short code: " + shortCode);
+    }
+}